+++ R/verbatim_popup.R
@@ -31,7 +31,14 @@ verbatim_popup_ui <- function(id, button_label, ...) {
     shiny::singleton(
       shiny::tags$head(shiny::includeScript(system.file("js/verbatim_popup.js", package = "teal.widgets")))
     ),
-    shiny::actionButton(ns("button"), label = button_label, ...)
+    shiny::actionButton(
+      ns("button"),
+      label =  div(
+        button_label,
+        uiOutput(ns("disable_controller"), inline = TRUE) # this is dummy output - see the comment in srv
+      ),
+      ...
+    )
   )
 }
 
@@ -53,7 +60,7 @@ verbatim_popup_srv <- function(id, verbatim_content, title, style = FALSE, disab
   checkmate::assert_class(disabled, classes = "reactive")
   moduleServer(id, function(input, output, session) {
     ns <- session$ns
-    disabled_flag_observer(disabled, "button")
+    disabled_flag_observer(disabled, "button", output)
     modal_content <- format_content(verbatim_content, style)
     button_click_observer(
       click_event = shiny::reactive(input$button),
@@ -73,17 +80,20 @@ verbatim_popup_srv <- function(id, verbatim_content, title, style = FALSE, disab
 #' @keywords internal
 #' @param disabled_flag (`reactive`) containing the flag
 #' @param button_id (`character(1)`) the id of the controlled button
-disabled_flag_observer <- function(disabled_flag, button_id) {
-  shiny::observeEvent(
-    disabled_flag(),
-    handlerExpr = {
-      if (disabled_flag()) {
-        shinyjs::disable(button_id)
-      } else {
-        shinyjs::enable(button_id)
-      }
+#' @param output (`shinyoutput`) needed to render output.
+disabled_flag_observer <- function(disabled_flag, button_id, output) {
+  # In normal case we could enable/disable using observeEvent(disabled_flag(), ...)
+  #  but observeEvent doesn't care whether output is displayed or not
+  #  This means that if we want to prevent calulation of disabled_flag for hidden
+  #  output, we need to use renderUI which triggers when the output is shown.
+  output$disable_controller <- shiny::renderUI({
+    if (disabled_flag()) {
+      shinyjs::disable(button_id)
+    } else {
+      shinyjs::enable(button_id)
     }
-  )
+    NULL
+  })
 }
 
 #' Creates a `shiny` observer handling button clicks.